feat(miniDistrictRankedCard): create a special componant to display district by rank.
