+++ app/home/page.tsx
@@ -1,8 +1,8 @@
 import { getTopDistricts, getlastReviews } from "./action";
-import Link from "next/link";
 import ReviewsCarrousel from "@/components/ui/Carrousel/ReviewsCarrousel";
 import SearchDistrict from "@/components/ui/SearchDistrict/SearchDistrict";
 import RankedDistrictCard from "@/components/ui/RankedDistrictCard/RankedDistrictCard";
+import MiniRankedDistrictCard from "@/components/ui/RankedDistrictCard/MiniRankedDistrictCard";
 import Modal from "@/components/core/modal/Modal";
 import iconDict from "@/modules/utils/iconDict";
 import criteriasDict from "@/modules/utils/criteriasDict";
@@ -83,28 +83,8 @@ export default async function Home() {
         <h3>Top 10 communities</h3>
         <div className="grid gap-4 grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
           {topDistricts.map((district) => {
-            const renderRankedColor = () => {
-              if (district.rank === 1)
-                return "bg-gradient-to-r from-yellow-500 via-yellow-300 to-yellow-500";
-              else if (district.rank === 2)
-                return "bg-gradient-to-r from-slate-500 via-slate-300 to-slate-500";
-              else if (district.rank === 3)
-                return "bg-gradient-to-r from-orange-800 via-orange-300 to-orange-800";
-              else return "bg-neutral";
-            };
             return (
-              <Link
-                href={`district/${district.id}`}
-                className={`flex justify-between ${renderRankedColor()} p-4 rounded-2xl shadow-md hover:bg-red`}
-                key={district.id}
-              >
-                <div className="flex gap-2 text-primary">
-                  <span className="!font-black">#{district.rank}</span>
-                  <span className="!font-bold">{district.name}</span>-{" "}
-                  {district.sector}
-                </div>
-                <p className="!font-black">{district.average_rating}</p>
-              </Link>
+              <MiniRankedDistrictCard key={district.id} district={district} />
             );
           })}
         </div>
